requirements.txt
streamlit
streamlit-lottie
python-docx
requests
langchain-groq
pymongo
python-dotenv

